// backend/src/common/enums/email.enum.ts
export enum EmailType {
  // Employee Registration & Onboarding
  EMPLOYEE_WELCOME = 'EMPLOYEE_WELCOME',
  EMPLOYEE_REGISTRATION_APPROVED = 'EMPLOYEE_REGISTRATION_APPROVED',
  EMPLOYEE_CREDENTIALS = 'EMPLOYEE_CREDENTIALS',
  EMPLOYEE_PROFILE_VERIFIED = 'EMPLOYEE_PROFILE_VERIFIED',
  
  // Manager Notifications
  MANAGER_REGISTRATION_REQUEST = 'MANAGER_REGISTRATION_REQUEST',
  MANAGER_REGISTRATION_APPROVED = 'MANAGER_REGISTRATION_APPROVED',
  MANAGER_REGISTRATION_REJECTED = 'MANAGER_REGISTRATION_REJECTED',
  
  // System Notifications
  PASSWORD_RESET = 'PASSWORD_RESET',
  ACCOUNT_LOCKED = 'ACCOUNT_LOCKED',
  SECURITY_ALERT = 'SECURITY_ALERT',
  
  // Workflow & Approvals
  WORKFLOW_APPROVAL_REQUEST = 'WORKFLOW_APPROVAL_REQUEST',
  WORKFLOW_APPROVED = 'WORKFLOW_APPROVED',
  WORKFLOW_REJECTED = 'WORKFLOW_REJECTED',
  
  // Payroll & Finance
  PAYSLIP_READY = 'PAYSLIP_READY',
  PAYROLL_PROCESSED = 'PAYROLL_PROCESSED',
  LOAN_APPLICATION_STATUS = 'LOAN_APPLICATION_STATUS',
  
  // Leave & Attendance
  LEAVE_APPLICATION_SUBMITTED = 'LEAVE_APPLICATION_SUBMITTED',
  LEAVE_APPROVED = 'LEAVE_APPROVED',
  LEAVE_REJECTED = 'LEAVE_REJECTED',
  ATTENDANCE_REMINDER = 'ATTENDANCE_REMINDER',
  
  // Training & Development
  TRAINING_INVITATION = 'TRAINING_INVITATION',
  TRAINING_REMINDER = 'TRAINING_REMINDER',
  TRAINING_COMPLETED = 'TRAINING_COMPLETED',
  
  // Performance Management
  PERFORMANCE_REVIEW_INVITATION = 'PERFORMANCE_REVIEW_INVITATION',
  PERFORMANCE_REVIEW_REMINDER = 'PERFORMANCE_REVIEW_REMINDER',
  GOAL_SETTING_REMINDER = 'GOAL_SETTING_REMINDER'
}

export enum EmailPriority {
  HIGH = 'high',
  NORMAL = 'normal',
  LOW = 'low'
}

export interface EmailTemplateData {
  [key: string]: any;
}

export interface EmailAttachment {
  filename: string;
  content: Buffer;
  contentType: string;
}